Move New swap page to cow-react folder
