+++ src/cow-react/pages/NewSwap/helpers/tokenViewAmount.ts
