+++ src/cow-react/pages/NewSwap/hooks/useSwapCurrenciesAmounts.ts
