+++ src/cow-react/pages/NewSwap/pureComponents/CurrencyArrowSeparator/__fixtures__/CurrencyArrowSeparator.tsx
@@ -1,0 +1,16 @@
+import {
+  CurrencyArrowSeparator,
+  CurrencyArrowSeparatorProps,
+} from 'cow-react/pages/NewSwap/pureComponents/CurrencyArrowSeparator'
+
+const defaultProps: CurrencyArrowSeparatorProps = {
+  isLoading: false,
+  withRecipient: false,
+  onSwitchTokens() {
+    //
+  },
+}
+
+export default {
+  Default: <CurrencyArrowSeparator {...defaultProps} />,
+}
