+++ src/cow-react/pages/NewSwap/pureComponents/CurrencyArrowSeparator/index.tsx
