+++ src/cow-react/pages/NewSwap/pureComponents/CurrencyArrowSeparator/style-mixins.tsx
