+++ src/cow-react/pages/NewSwap/pureComponents/CurrencyArrowSeparator/styled.tsx
