+++ src/cow-react/pages/NewSwap/pureComponents/CurrencyInputPanel/styled.tsx
