+++ src/cow-react/pages/NewSwap/pureComponents/CurrencySelectButton/index.tsx
