+++ src/cow-react/pages/NewSwap/pureComponents/CurrencySelectButton/styled.tsx
