+++ src/cow-react/pages/NewSwap/pureComponents/ReceiveAmount/styled.tsx
