+++ src/cow-react/pages/NewSwap/pureComponents/ReceiveAmountInfo/styled.tsx
