+++ src/cow-react/pages/NewSwap/pureComponents/TradeRates/styled.tsx
