+++ src/custom/pages/App/index.tsx
@@ -19,7 +19,7 @@ import { Loading } from 'components/FlashingLoading'
 // Sync routes
 import Account from 'pages/Account'
 import Swap from 'pages/Swap'
-import { NewSwapPage } from 'pages/NewSwap'
+import { NewSwapPage } from 'cow-react/pages/NewSwap'
 
 const SENTRY_DSN = process.env.REACT_APP_SENTRY_DSN
 const SENTRY_TRACES_SAMPLE_RATE = process.env.REACT_APP_SENTRY_TRACES_SAMPLE_RATE
