--- src/custom/pages/NewSwap/pureComponents/CurrencyArrowSeparator/__fixtures__/CurrencyArrowSeparator.tsx
@@ -1,16 +1,0 @@
-import {
-  CurrencyArrowSeparator,
-  CurrencyArrowSeparatorProps,
-} from 'pages/NewSwap/pureComponents/CurrencyArrowSeparator'
-
-const defaultProps: CurrencyArrowSeparatorProps = {
-  isLoading: false,
-  withRecipient: false,
-  onSwitchTokens() {
-    //
-  },
-}
-
-export default {
-  Default: <CurrencyArrowSeparator {...defaultProps} />,
-}
